Fix MSVC compiler warnings
This fixes warnings from the MSVC compiler (Version 19.29.30136 for x64) when compiling a 64-bit app.

warning C4244: '=': conversion from 'int' to 'wxTextAttrDimensionFlags', possible loss of data

warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data

+++ include/wx/choicebk.h
@@ -86,7 +86,7 @@ protected:
 
     void UpdateSelectedPage(size_t newsel) wxOVERRIDE
     {
-        GetChoiceCtrl()->Select(newsel);
+        GetChoiceCtrl()->Select(static_cast<int>(newsel));
     }
 
     wxBookCtrlEvent* CreatePageChangingEvent() const wxOVERRIDE;
